@@ -1,52 +1,56 @@
 {
-  "name": "rfp-ng2",
-  "version": "2.0.1",
-  "license": "MIT",
-  "angular-cli": {},
-  "scripts": {
-    "ng": "ng",
-    "start": "ng serve",
-    "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e"
-  },
-  "private": true,
-  "dependencies": {
-    "@angular/common": "^2.4.0",
-    "@angular/compiler": "^2.4.0",
-    "@angular/core": "^2.4.0",
-    "@angular/forms": "^2.4.0",
-    "@angular/http": "^2.4.0",
-    "@angular/platform-browser": "^2.4.0",
-    "@angular/platform-browser-dynamic": "^2.4.0",
-    "@angular/router": "^3.4.0",
-    "bootstrap": "^3.3.7",
-    "core-js": "^2.4.1",
-    "font-awesome": "^4.7.0",
-    "lodash": "^4.17.4",
-    "ng2-bootstrap": "^1.3.3",
-    "primeng": "^2.0.0",
-    "rxjs": "^5.0.1",
-    "ts-helpers": "^1.1.1",
-    "zone.js": "^0.7.2"
-  },
-  "devDependencies": {
-    "@angular/cli": "1.0.0-beta.31",
-    "@angular/compiler-cli": "^2.4.0",
-    "@types/jasmine": "2.5.38",
-    "@types/lodash": "^4.14.52",
-    "@types/node": "^6.0.42",
-    "codelyzer": "~2.0.0-beta.1",
-    "jasmine-core": "2.5.2",
-    "jasmine-spec-reporter": "2.5.0",
-    "karma": "1.2.0",
-    "karma-chrome-launcher": "^2.0.0",
-    "karma-cli": "^1.0.1",
-    "karma-jasmine": "^1.0.2",
-    "karma-coverage-istanbul-reporter": "^0.2.0",
-    "protractor": "~5.1.0",
-    "ts-node": "1.2.1",
-    "tslint": "^4.3.0",
-    "typescript": "~2.0.0"
-  }
+    "name": "rfp-ng2",
+    "version": "2.0.1",
+    "license": "MIT",
+    "angular-cli": {},
+    "scripts": {
+        "ng": "ng",
+        "start": "ng serve",
+        "build": "ng build",
+        "test": "ng test",
+        "lint": "ng lint",
+        "e2e": "ng e2e"
+    },
+    "private": true,
+    "dependencies": {
+        "@angular/animations": "^4.0.1",
+        "@angular/common": "^4.0.1",
+        "@angular/compiler": "^4.0.1",
+        "@angular/core": "^4.0.1",
+        "@angular/forms": "^4.0.1",
+        "@angular/http": "^4.0.1",
+        "@angular/platform-browser": "^4.0.1",
+        "@angular/platform-browser-dynamic": "^4.0.1",
+        "@angular/router": "^4.0.1",
+        "angular2-jwt": "^0.2.0",
+        "auth0-lock": "^10.14.0",
+        "bootstrap": "^4.0.0-alpha.6",
+        "core-js": "^2.4.1",
+        "font-awesome": "^4.7.0",
+        "lodash": "^4.17.4",
+        "ng2-bootstrap": "^1.6.1",
+        "primeng": "^4.0.0-rc.2",
+        "rxjs": "^5.2.0",
+        "zone.js": "^0.8.5"
+    },
+    "devDependencies": {
+        "@angular/cli": "^1.0.0",
+        "@angular/compiler-cli": "^4.0.1",
+        "@types/jasmine": "2.5.47",
+        "@types/lodash": "^4.14.61",
+        "@types/node": "^7.0.12",
+        "codelyzer": "~2.1.1",
+        "jasmine-core": "2.5.2",
+        "jasmine-spec-reporter": "3.2.0",
+        "karma": "1.5.0",
+        "karma-chrome-launcher": "^2.0.0",
+        "karma-cli": "^1.0.1",
+        "karma-coverage-istanbul-reporter": "^1.0.0",
+        "karma-jasmine": "^1.1.0",
+        "karma-jasmine-html-reporter": "^0.2.2",
+        "protractor": "~5.1.1",
+        "ts-node": "3.0.2",
+        "tslint": "^4.5.0",
+        "typescript": "~2.2.2"
+    }
 }
