@@ -1,7 +1,7 @@
 import {BrowserModule} from '@angular/platform-browser';
 import {NgModule} from '@angular/core';
 import {FormsModule} from '@angular/forms';
-import {HttpModule} from '@angular/http';
+import {HttpModule, RequestOptions, Http} from '@angular/http';
 import {routing, appRoutingProviders}  from './app.routing';
 
 import {AppComponent} from './app.component';
@@ -14,6 +14,9 @@ import {PositionModule} from "./position/position.module";
 import {SubPositionModule} from "./sub-position/sub-position.module";
 import {GenericModule} from "./generic/generic.module";
 import {Logger} from "./generic/logger.service";
+import {AuthService} from "./auth/auth.service";
+import {authFactory} from "./auth/auth.factory";
+import {AuthHttp} from "angular2-jwt";
 
 @NgModule({
     declarations: [
@@ -35,7 +38,13 @@ import {Logger} from "./generic/logger.service";
     ],
     providers: [
         appRoutingProviders,
-        Logger
+        AuthService,
+        Logger,
+        {
+            provide: AuthHttp,
+            useFactory: authFactory,
+            deps: [Http, RequestOptions]
+        }
     ],
     bootstrap: [AppComponent]
 })
