@@ -1,15 +1,15 @@
 import {Injectable} from '@angular/core';
-import {Http} from '@angular/http';
 
 import {Logger} from "../generic/logger.service"
 import {GenericService} from "../generic/generic.service";
 import {Invoice} from "./invoice";
+import {AuthHttp} from "angular2-jwt";
 
 @Injectable()
 export class InvoiceService extends GenericService<Invoice> {
 
-    constructor(_http: Http, _logger: Logger) {
-        super(_http, _logger);
+    constructor(http: AuthHttp, logger: Logger) {
+        super(http, logger);
     }
 
     protected modelName() { return 'invoice'; }
