@@ -1,5 +1,5 @@
 import {Injectable} from '@angular/core';
-import {Http, Response} from '@angular/http';
+import {Response} from '@angular/http';
 import {Observable} from 'rxjs/Observable';
 import 'rxjs/add/operator/do'
 import 'rxjs/add/operator/map'
@@ -8,12 +8,13 @@ import 'rxjs/add/operator/catch'
 import {Logger} from "../generic/logger.service"
 import {GenericService} from "../generic/generic.service";
 import {Position} from "./position";
+import {AuthHttp} from "angular2-jwt";
 
 @Injectable()
 export class PositionService extends GenericService<Position> {
 
-    constructor(_http: Http, _logger: Logger) {
-        super(_http, _logger);
+    constructor(http: AuthHttp, logger: Logger) {
+        super(http, logger);
     }
 
     protected modelName() { return 'position'; }
@@ -25,9 +26,9 @@ export class PositionService extends GenericService<Position> {
         if(storeId)
             url += `/${storeId}`;
 
-        return this._http.get(url)
+        return this.http.get(url)
             .map((res:Response) => <Position> res.json())
-            .do(data => this._logger.debug(`GET /search ${this.modelName()} -> ${JSON.stringify(data)}`))
+            .do(data => this.logger.debug(`GET /search ${this.modelName()} -> ${JSON.stringify(data)}`))
             .catch(this.handleError);
     }
 
@@ -35,9 +36,9 @@ export class PositionService extends GenericService<Position> {
 
         var url = `${this.url}/invoice/${invoiceId}`;
 
-        return this._http.get(url)
+        return this.http.get(url)
             .map((res:Response) => <Position[]> res.json())
-            .do(data => this._logger.debug(`GET ${this.modelName()}(s) -> ${JSON.stringify(data)}`))
+            .do(data => this.logger.debug(`GET ${this.modelName()}(s) -> ${JSON.stringify(data)}`))
             .catch(this.handleError);
     }
 }
