@@ -1,5 +1,5 @@
 import {Injectable} from '@angular/core';
-import {Http, Response} from '@angular/http';
+import {Response} from '@angular/http';
 import {Observable} from 'rxjs/Observable';
 import 'rxjs/add/operator/do'
 import 'rxjs/add/operator/map'
@@ -8,12 +8,13 @@ import 'rxjs/add/operator/catch'
 import {Logger} from "../generic/logger.service";
 import {GenericService} from "../generic/generic.service";
 import {Product} from "./product";
+import {AuthHttp} from "angular2-jwt";
 
 @Injectable()
 export class ProductService extends GenericService<Product> {
 
-    constructor(_http:Http, _logger:Logger) {
-        super(_http, _logger);
+    constructor(http: AuthHttp, logger: Logger) {
+        super(http, logger);
     }
 
     protected modelName() {
@@ -26,27 +27,27 @@ export class ProductService extends GenericService<Product> {
         if (limit)
             search_url += `/${limit}`;
 
-        return this._http.get(search_url)
+        return this.http.get(search_url)
             .map((res:Response) => <Product[]> res.json())
-            .do(data => this._logger.debug(`GET /search ${this.modelName()} -> ${JSON.stringify(data)}`))
+            .do(data => this.logger.debug(`GET /search ${this.modelName()} -> ${JSON.stringify(data)}`))
             .catch(this.handleError);
     }
 
     show_children(id:string):Observable<Product> {
         var url = `${this.url}/${id}/show_children`;
 
-        return this._http.get(url)
+        return this.http.get(url)
             .map((res:Response) => <Product> res.json())
-            .do(data => this._logger.debug(`GET /show_child ${this.modelName()} -> ${JSON.stringify(data)}`))
+            .do(data => this.logger.debug(`GET /show_child ${this.modelName()} -> ${JSON.stringify(data)}`))
             .catch(this.handleError);
     }
 
     add_child(fatherId:string, childId:string):Observable<Product> {
         var url = `${this.url}/${fatherId}/add_child/${childId}`;
 
-        return this._http.get(url)
+        return this.http.get(url)
             .map((res:Response) => <Product> res.json())
-            .do(data => this._logger.debug(`GET /add_child ${this.modelName()} -> ${JSON.stringify(data)}`))
+            .do(data => this.logger.debug(`GET /add_child ${this.modelName()} -> ${JSON.stringify(data)}`))
             .catch(this.handleError);
     }
 
@@ -54,9 +55,9 @@ export class ProductService extends GenericService<Product> {
 
         var url = `${this.url}/${fatherId}/remove_child/${childId}`;
 
-        return this._http.get(url)
+        return this.http.get(url)
             .map((res:Response) => <Product> res.json())
-            .do(data => this._logger.debug(`GET /remove_child ${this.modelName()} -> ${JSON.stringify(data)}`))
+            .do(data => this.logger.debug(`GET /remove_child ${this.modelName()} -> ${JSON.stringify(data)}`))
             .catch(this.handleError);
     }
 }
