@@ -1,5 +1,4 @@
 import {Injectable} from '@angular/core';
-import {Http} from '@angular/http';
 import 'rxjs/add/operator/do'
 import 'rxjs/add/operator/map'
 import 'rxjs/add/operator/catch'
@@ -7,12 +6,13 @@ import 'rxjs/add/operator/catch'
 import {Logger} from "../generic/logger.service";
 import {GenericService} from "../generic/generic.service";
 import {SubPosition} from "./sub-position";
+import {AuthHttp} from "angular2-jwt";
 
 @Injectable()
 export class SubPositionService extends GenericService<SubPosition> {
 
-    constructor(_http: Http, _logger: Logger) {
-        super(_http, _logger);
+    constructor(http: AuthHttp, logger: Logger) {
+        super(http, logger);
     }
 
     protected modelName() { return 'sub-position'; }
